feat: fix subscription request after update on database

+++ src/main/java/com/iaraapi/dto/response/PaymentResponse.java
@@ -3,13 +3,13 @@ package com.iaraapi.dto.response;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
-import java.util.UUID;
 
 @Getter
 @AllArgsConstructor
 public class PaymentResponse {
-    private Double total;
+    private BigDecimal total;
     private LocalDateTime expiresOn;
     private String paymentMethod;
     private String subscription;
